// React/app/GerAdv_TS/TipoEnderecoSistema/Interfaces/interface.TipoEnderecoSistema.ts
// 100% auto-generated by Source Genesys WebApi Front & Back Creator
// copyright © 2000-2025 Menphis - Sistemas Inteligentes
// This file is part of the Source Genesys project                     
'use client';

export interface ITipoEnderecoSistema {
// 202501251
    id: number;
 
	nome: string,
}

export interface ITipoEnderecoSistemaFormProps {
  id: number;
  onClose: () => void;
  onSuccess: (registro?: any) => void;
  onError: () => void;
}

export interface ITipoEnderecoSistemaIncProps {
    id: number;
    onClose: () => void;
    onSuccess: (registro?: any) => void;
    onError: () => void;
}